Autostart the registration/authentication
The gesture (buttonpress) is no longer required to start the WebAuthn
session with the browser. This is a revert of the original behavior. But
the button remained on the page for users of an old unsupportive
browser.

See https://www.pivotaltracker.com/story/show/184695105
