+++ public/typescript/ui/component/AuthenticationContainer.tsx
@@ -1,8 +1,10 @@
-import React, { FC } from 'react';
+import React, { FC, useEffect } from 'react';
 import { RequestInformation, SerializedPublicKeyCredentialRequestOptions, TranslationString } from '../../model';
 import { useAppReducer, useAuthenticationEffect, useClickable } from '../hook';
 import { useVerifyPublicKeyCredentials } from '../hook/useVerifyPublicKeyCredentials';
 import { App } from './App';
+import { startAuthentication } from '../../observable/startAuthentication';
+import { Subscription } from 'rxjs';
 
 export interface AuthenticationContainerProps {
   publicKeyOptions: SerializedPublicKeyCredentialRequestOptions;
@@ -17,5 +19,11 @@ export const AuthenticationContainer: FC<AuthenticationContainerProps> = ({ t, r
   const [click, clicked] = useClickable();
   const verify = useVerifyPublicKeyCredentials(responseUrl);
   const onStart = useAuthenticationEffect(dispatch, publicKeyOptions, verify, clicked);
-  return <App started={started} startMessage="authentication.start_button" message={message} errorInfo={errorInfo} requestInformation={requestInformation} t={t} onClick={click} onStart={onStart} />;
+  useEffect(() => {
+    const subscription: Subscription = startAuthentication(dispatch, publicKeyOptions, verify, clicked);
+    return () => {
+      subscription.unsubscribe();
+    };
+  },        [dispatch, publicKeyOptions, verify, clicked]);
+  return <App started={started} startMessage="authentication.start_button" message={message} errorInfo={errorInfo} requestInformation={requestInformation} t={t} onClick={click} onStart={onStart}/>;
 };
